@@ -1,23 +1,22 @@
 import 'dart:async';
 
-import 'package:flutter/cupertino.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:trus_app/features/auth/controller/auth_controller.dart';
 import 'package:trus_app/features/auth/repository/auth_repository.dart';
-import 'package:trus_app/features/general/crud_operations.dart';
 import 'package:trus_app/features/general/global_variables_controller.dart';
-import 'package:trus_app/features/mixin/achievement_controller_mixin.dart';
-import 'package:trus_app/features/mixin/boolean_controller_mixin.dart';
-import 'package:trus_app/features/mixin/date_controller_mixin.dart';
 import 'package:trus_app/features/mixin/string_controller_mixin.dart';
 import 'package:trus_app/features/mixin/view_controller_mixin.dart';
 import 'package:trus_app/features/user/widget/i_user_key.dart';
 import 'package:trus_app/models/api/auth/user_api_model.dart';
+import 'package:trus_app/models/api/auth/user_setup.dart';
 import 'package:trus_app/models/api/auth/user_team_role_api_model.dart';
+import 'package:trus_app/models/api/interfaces/confirm_to_string.dart';
+import 'package:trus_app/models/api/interfaces/impl/confirm_to_string_impl.dart';
 
 import '../../../common/repository/exception/internal_snackbar_exception.dart';
 import '../../../common/utils/field_validator.dart';
 import '../../../models/api/player/player_api_model.dart';
+import '../../general/confirm_operations.dart';
 import '../../general/read_operations.dart';
 import '../../mixin/dropdown_controller_mixin.dart';
 
@@ -27,17 +26,11 @@ final userControllerProvider = Provider((ref) {
 });
 
 class UserController
-    with
-        DropdownControllerMixin,
-        StringControllerMixin,
-        DateControllerMixin,
-        BooleanControllerMixin,
-        ViewControllerMixin,
-        AchievementControllerMixin
-    implements CrudOperations, ReadOperations, IUserKey {
+    with DropdownControllerMixin, StringControllerMixin, ViewControllerMixin
+    implements ReadOperations, IUserKey, ConfirmOperations {
   final AuthRepository authApiService;
   final Ref ref;
-  late UserApiModel userApiModel;
+  late UserSetup userSetup;
 
   UserController({
     required this.authApiService,
@@ -46,24 +39,23 @@ class UserController
 
   void loadViewUser() {
     int appTeamId = ref.read(globalVariablesControllerProvider).appTeam!.id;
-    initViewFields(userApiModel.mail ?? "", emailKey());
-    initViewFields(userApiModel.name ?? "", nameKey());
-    initViewFields(
-        getPlayerFromUserTeamRole(
-            userApiModel.getCurrentUserTeamRole(appTeamId)),
+    initViewFields(userSetup.currentUser.mail ?? "", emailKey());
+    initViewFields(userSetup.currentUser.name ?? "", nameKey());
+    initDropdown(userSetup.primaryPlayer, userSetup.eligiblePlayersToPairWith,
         playerKey());
     initViewFields(
-        getRoleFromUserTeamRole(userApiModel.getCurrentUserTeamRole(appTeamId)),
+        getRoleFromUserTeamRole(
+            userSetup.currentUser.getCurrentUserTeamRole(appTeamId)),
         roleKey());
-    initViewFields(
-        userApiModel.getDescriptionOfOtherRoles(appTeamId), otherRolesKey());
+    initViewFields(userSetup.currentUser.getDescriptionOfOtherRoles(appTeamId),
+        otherRolesKey());
   }
 
-  String getPlayerFromUserTeamRole(UserTeamRoleApiModel? teamRole) {
+  PlayerApiModel? getPlayerFromUserTeamRole(UserTeamRoleApiModel? teamRole) {
     if (teamRole == null) {
-      return "";
+      return null;
     }
-    return teamRole.playerToString();
+    return teamRole.player;
   }
 
   String getRoleFromUserTeamRole(UserTeamRoleApiModel? teamRole) {
@@ -74,7 +66,7 @@ class UserController
   }
 
   Future<void> setupUser() async {
-    userApiModel = await _setupUser();
+    userSetup = await _setupUser();
   }
 
   Future<void> viewUser() async {
@@ -87,30 +79,12 @@ class UserController
     return errorText.isEmpty;
   }
 
-  Future<UserApiModel> _setupUser() async {
-    final user = await authApiService.getCurrentUserData();
-    if (user == null) {
-      throw Exception("Uživatel nebyl načten.");
-    }
+  Future<UserSetup> _setupUser() async {
+    final user = await authApiService.getUserSetup();
     return user;
   }
 
-  @override
-  Future<PlayerApiModel?> addModel() async {
-    return null;
-  }
-
-  @override
-  Future<String> deleteModel(int id) async {
-    final bool = await authApiService.deleteAccount();
-    if (bool) {
-      return "Účet smazán";
-    }
-    return "Chyba při smazání účtu";
-  }
-
   Future<void> changeWritePermissions(
-    BuildContext context,
     UserApiModel user,
   ) async {
     if (user.teamRoles == null || user.teamRoles!.isEmpty) {
@@ -121,12 +95,8 @@ class UserController
       throw InternalSnackBarException("Nemůžeš změnit práva sám sobě");
     }
     String newRole = user.teamRoles![0].getOppositeRole();
-    await authApiService.setUserWritePermissions(user.teamRoles![0].id!, newRole);
-  }
-
-  @override
-  Future<String?> editModel(int id) async {
-    return null;
+    await authApiService.setUserWritePermissions(
+        user.teamRoles![0].id!, newRole);
   }
 
   @override
@@ -157,5 +127,18 @@ class UserController
   @override
   String nameKey() {
     return "user_name";
+  }
+
+  @override
+  Future<ConfirmToString> addModel(int id) async {
+    PlayerApiModel playerApiModel = dropdownValues[playerKey()] as PlayerApiModel;
+    await authApiService
+        .setUserPlayerId(dropdownValues[playerKey()] as PlayerApiModel);
+    if(playerApiModel.id == 0) {
+      ref.read(globalVariablesControllerProvider).setPlayerApiModel(null);
+      return ConfirmToStringImpl("Z profilu odebrán hráč");
+    }
+    ref.read(globalVariablesControllerProvider).setPlayerApiModel(playerApiModel);
+    return ConfirmToStringImpl("Do profilu přidán hráč ${playerApiModel.name}");
   }
 }
