@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:trus_app/features/fine/match/controller/fine_player_controller.dart';
-import 'package:trus_app/features/fine/match/screens/fine_match_screen.dart';
 import 'package:trus_app/models/api/match/match_api_model.dart';
 
 import '../../../../common/widgets/builder/add_builder.dart';
@@ -10,6 +9,7 @@ import '../../../../common/widgets/button/confirm_button.dart';
 import '../../../../common/widgets/screen/custom_consumer_stateful_widget.dart';
 import '../../../../models/api/player/player_api_model.dart';
 import '../../../main/screen_controller.dart';
+import 'fine_match_screen.dart';
 
 class FinePlayerScreen extends CustomConsumerStatefulWidget {
   static const String id = "fine-player-screen";
