@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:trus_app/common/widgets/rows/row_switch.dart';
+import 'package:trus_app/common/widgets/rows/row_warning_text.dart';
+import 'package:trus_app/features/fine/controller/fine_edit_notifier.dart';
+import 'package:trus_app/features/fine/controller/fine_notifier.dart';
 import 'package:trus_app/features/fine/screens/fine_screen.dart';
-import 'package:trus_app/models/api/fine_api_model.dart';
 
-import '../../../common/widgets/builder/column_future_builder.dart';
-import '../../../common/widgets/button/crud_button.dart';
-import '../../../common/widgets/rows/crud/row_switch_stream.dart';
-import '../../../common/widgets/rows/crud/row_text_field_stream.dart';
+import '../../../common/widgets/button/simple_crud_button.dart';
+import '../../../common/widgets/notifier/loader/loading_overlay.dart';
+import '../../../common/widgets/rows/row_text_field.dart';
 import '../../../common/widgets/screen/custom_consumer_stateful_widget.dart';
 import '../../../models/enum/crud.dart';
-import '../../main/screen_controller.dart';
-import '../controller/fine_controller.dart';
 
 class EditFineScreen extends CustomConsumerStatefulWidget {
   static const String id = "edit-fine-screen";
@@ -26,90 +26,69 @@ class EditFineScreen extends CustomConsumerStatefulWidget {
 class _EditFineScreenState extends ConsumerState<EditFineScreen> {
   @override
   Widget build(BuildContext context) {
-    if (ref.read(screenControllerProvider).isScreenFocused(EditFineScreen.id)) {
-      FineApiModel fine = ref.watch(screenControllerProvider).fineModel;
-      const double padding = 8.0;
-      final size =
-          MediaQueryData.fromView(WidgetsBinding.instance.window).size;
-      return ColumnFutureBuilder(
-        loadModelFuture: ref.watch(fineControllerProvider).fine(fine),
-        columns: [
-          RowTextFieldStream(
-            key: const ValueKey('fine_name_field'),
-            size: size,
-            labelText: "název",
-            textFieldText: "Název pokuty:",
-            padding: padding,
-            stringControllerMixin: ref.watch(fineControllerProvider),
-            hashKey: ref.read(fineControllerProvider).nameKey,
-          ),
-          const SizedBox(height: 10),
-          RowTextFieldStream(
-            key: const ValueKey('fine_amount_field'),
-            size: size,
-            labelText: "v Kč",
-            textFieldText: "Výše pokuty:",
-            padding: padding,
-            stringControllerMixin: ref.watch(fineControllerProvider),
-            hashKey: ref.read(fineControllerProvider).amountKey,
-            number: true,
-          ),
-          const SizedBox(height: 10),
-          RowSwitchStream(
-              key: const ValueKey('fine_switch_field'),
-              size: size,
-              padding: padding,
-              textFieldText: "Pouze pro nově udělené pokuty?",
-              booleanControllerMixin: ref.watch(fineControllerProvider),
-              hashKey: ref.read(fineControllerProvider).inactiveKey),
-          const SizedBox(
-            height: 20,
-          ),
-          Row(
+    final fine = ref.watch(fineNotifierProvider).selectedFine;
+    final state = ref.watch(fineEditProvider(fine));
+    final notifier = ref.read(fineEditProvider(fine).notifier);
+    return LoadingOverlay(
+        state: state,
+        onClearError: notifier.clearErrorMessage,
+        child: Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Column(
             children: [
-              SizedBox(
-                  width: size.width / 6,
-                  child: const Icon(
-                    Icons.warning,
-                    color: Colors.black,
-                  )),
-              SizedBox(
-                  width: size.width / 1.3,
-                  child: const Text(
+              RowTextField(
+                label: "Název",
+                textFieldText: "Název pokuty:",
+                value: state.name,
+                onChanged: notifier.setName,
+                error: state.errors["name"],
+              ),
+              const SizedBox(height: 10),
+              RowTextField(
+                label: "v Kč",
+                textFieldText: "Výše pokuty:",
+                value: state.amount,
+                number: true,
+                onChanged: notifier.setAmount,
+                error: state.errors["amount"],
+              ),
+              RowSwitch(
+                textFieldText: "Pouze pro nově udělené pokuty?",
+                value: state.inactive,
+                onChanged: notifier.setInactive,
+              ),
+              const SizedBox(
+                height: 20,
+              ),
+              const RowWarningText(
+                  text:
                       "Dbejte na to, že pokud je tento přepínač vypnutý, tak se změní všechny pokuty i retrospektivně - tedy částka či jméno se změní i pro již udělené pokuty z předchozích zápasů.\n"
-                      "Pokud si přejete udělat změny pouze pro nové pokuty, přepněte do přepínač do true"))
+                      "Pokud si přejete udělat změny pouze pro nové pokuty, přepněte do přepínač do true"),
+              const SizedBox(height: 10),
+              const SizedBox(height: 10),
+              SimpleCrudButton(
+                onPressed: () async => await notifier.submit(
+                    "Ukládám pokutu...",
+                    "Pokuta byla úspěšně uložena",
+                    FineScreen.id,
+                    Crud.update,
+                    fineNotifierProvider),
+                text: "Potvrď změny",
+              ),
+              const SizedBox(height: 10),
+              SimpleCrudButton(
+                onPressed: () async => await notifier.submit(
+                    "Mažu pokutu...",
+                    "Pokutu byla úspěšně smazána",
+                    FineScreen.id,
+                    Crud.delete,
+                    fineNotifierProvider),
+                text: "Smaž pokutu",
+                deleteConfirmationText:
+                    "Opravdu chcete smazat pokutu ${state.model?.name}?",
+              ),
             ],
           ),
-          const SizedBox(height: 10),
-          const SizedBox(height: 10),
-          CrudButton(
-            key: const ValueKey('confirm_button'),
-            text: "Potvrď změny",
-            context: context,
-            crud: Crud.update,
-            crudOperations: ref.read(fineControllerProvider),
-            onOperationComplete: (id) {
-              ref.read(screenControllerProvider).changeFragment(FineScreen.id);
-            },
-            id: fine.id!,
-          ),
-          CrudButton(
-            key: const ValueKey('delete_button'),
-            text: "Smaž pokutu",
-            context: context,
-            crud: Crud.delete,
-            crudOperations: ref.read(fineControllerProvider),
-            onOperationComplete: (id) {
-              ref.read(screenControllerProvider).changeFragment(FineScreen.id);
-            },
-            id: fine.id!,
-            modelToString: fine,
-          ),
-        ],
-        loadingScreen: null,
-      );
-    } else {
-      return Container();
-    }
+        ));
   }
 }
