@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:trus_app/common/widgets/builder/models_cache_builder.dart';
+import 'package:trus_app/features/fine/controller/fine_notifier.dart';
 import 'package:trus_app/features/fine/screens/add_fine_screen.dart';
-import 'package:trus_app/features/fine/screens/edit_fine_screen.dart';
 
+import '../../../common/widgets/notifier/listview/model_to_string_listview.dart';
 import '../../../common/widgets/screen/custom_consumer_widget.dart';
-import '../../../models/api/fine_api_model.dart';
 import '../../main/screen_controller.dart';
-import '../controller/fine_controller.dart';
 
 class FineScreen extends CustomConsumerWidget {
   static const String id = "fine-screen";
@@ -18,35 +16,20 @@ class FineScreen extends CustomConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    if (ref.read(screenControllerProvider).isScreenFocused(FineScreen.id)) {
-      var provider = ref.watch(fineControllerProvider);
-      return Scaffold(
-          body: Padding(
-            padding: const EdgeInsets.only(top: 8.0),
-            child: ModelsCacheBuilder(
-              key: const ValueKey('fine_list'),
-              listSetup: provider.setupFineList(),
-              modelToStringListControllerMixin: provider,
-              hashKey: FineController.fineListId,
-              onPressed: (fine) => {
-                ref
-                    .read(screenControllerProvider)
-                    .setFine(fine as FineApiModel),
-                ref
-                    .read(screenControllerProvider)
-                    .changeFragment(EditFineScreen.id)
-              },
-            ),
-          ),
-          floatingActionButton: FloatingActionButton(
-            onPressed: () => ref
-                .read(screenControllerProvider)
-                .changeFragment(AddFineScreen.id),
-            elevation: 4.0,
-            child: const Icon(Icons.add),
-          ));
-    } else {
-      return Container();
-    }
+    return Scaffold(
+        body: Padding(
+          padding: const EdgeInsets.only(top: 8.0),
+          child: ModelToStringListview(
+              state: ref.watch(fineNotifierProvider),
+              notifier: ref.read(fineNotifierProvider.notifier)),
+        ),
+        floatingActionButton: FloatingActionButton(
+          key: const ValueKey('add_fine_floating_button'),
+          onPressed: () => ref
+              .read(screenControllerProvider)
+              .changeFragment(AddFineScreen.id),
+          elevation: 4.0,
+          child: const Icon(Icons.add),
+        ));
   }
 }
