@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:trus_app/features/season/controller/season_edit_notifier.dart';
 import 'package:trus_app/features/season/screens/season_screen.dart';
+import 'package:trus_app/models/enum/crud.dart';
 
-import '../../../common/widgets/builder/column_future_builder.dart';
-import '../../../common/widgets/button/crud_button.dart';
-import '../../../common/widgets/rows/crud/row_calendar_stream.dart';
-import '../../../common/widgets/rows/crud/row_text_field_stream.dart';
+import '../../../common/widgets/button/simple_crud_button.dart';
+import '../../../common/widgets/notifier/loader/loading_overlay.dart';
+import '../../../common/widgets/rows/row_date_picker.dart';
+import '../../../common/widgets/rows/row_text_field.dart';
 import '../../../common/widgets/screen/custom_consumer_stateful_widget.dart';
-import '../../../models/enum/crud.dart';
-import '../../main/screen_controller.dart';
-import '../controller/season_controller.dart';
+import '../controller/season_notifier.dart';
 
 class AddSeasonScreen extends CustomConsumerStatefulWidget {
   static const String id = "add-season-screen";
@@ -25,59 +25,49 @@ class AddSeasonScreen extends CustomConsumerStatefulWidget {
 class _AddSeasonScreenState extends ConsumerState<AddSeasonScreen> {
   @override
   Widget build(BuildContext context) {
-    if (ref
-        .read(screenControllerProvider)
-        .isScreenFocused(AddSeasonScreen.id)) {
-      const double padding = 8.0;
-      final size = MediaQueryData.fromView(WidgetsBinding.instance.window).size;
-      return ColumnFutureBuilder(
-        loadModelFuture: ref.watch(seasonControllerProvider).newSeason(),
-        columns: [
-          RowTextFieldStream(
-            key: const ValueKey('season_name_field'),
-            size: size,
-            labelText: "název",
-            padding: padding,
-            textFieldText: "Název sezony:",
-            stringControllerMixin: ref.watch(seasonControllerProvider),
-            hashKey: ref.read(seasonControllerProvider).nameKey,
+    final state = ref.watch(seasonEditProvider(null));
+    final notifier = ref.read(seasonEditProvider(null).notifier);
+    return LoadingOverlay(
+        state: state,
+        onClearError: notifier.clearErrorMessage,
+        child: Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Column(
+            children: [
+              RowTextField(
+                label: "Název",
+                textFieldText: "Název sezony:",
+                value: state.name,
+                onChanged: notifier.setName,
+                error: state.errors["name"],
+              ),
+              const SizedBox(height: 10),
+              RowDatePicker(
+                textFieldText: "Začátek sezony",
+                value: state.from,
+                onChanged: notifier.setFrom,
+                error: state.errors["fromDate"],
+              ),
+              const SizedBox(height: 10),
+              RowDatePicker(
+                textFieldText: "Konec sezony",
+                value: state.to,
+                onChanged: notifier.setTo,
+                error: state.errors["toDate"],
+              ),
+              const SizedBox(height: 10),
+              const SizedBox(height: 10),
+              SimpleCrudButton(
+                onPressed: () async => await notifier.submit(
+                    "Ukládám sezonu...",
+                    "Sezona byla úspěšně uložena",
+                    SeasonScreen.id,
+                    Crud.create,
+                    seasonNotifierProvider),
+                text: "Ulož pokutu",
+              ),
+            ],
           ),
-          const SizedBox(height: 10),
-          RowCalendarStream(
-            key: const ValueKey('season_start_date_field'),
-            size: size,
-            padding: padding,
-            textFieldText: "Začátek sezony:",
-            dateControllerMixin: ref.watch(seasonControllerProvider),
-            hashKey: ref.read(seasonControllerProvider).fromKey,
-          ),
-          const SizedBox(height: 10),
-          RowCalendarStream(
-            key: const ValueKey('season_end_date_field'),
-            size: size,
-            padding: padding,
-            textFieldText: "Konec sezony:",
-            dateControllerMixin: ref.watch(seasonControllerProvider),
-            hashKey: ref.read(seasonControllerProvider).toKey,
-          ),
-          const SizedBox(height: 10),
-          CrudButton(
-            key: const ValueKey('confirm_button'),
-            text: "Potvrď",
-            context: context,
-            crud: Crud.create,
-            crudOperations: ref.read(seasonControllerProvider),
-            onOperationComplete: (id) {
-              ref
-                  .read(screenControllerProvider)
-                  .changeFragment(SeasonScreen.id);
-            },
-          )
-        ],
-        loadingScreen: null,
-      );
-    } else {
-      return Container();
-    }
+        ));
   }
 }
