@@ -1,15 +1,14 @@
 import 'package:trus_app/models/api/achievement/achievement_player_detail.dart';
 import 'package:trus_app/models/api/football/football_player_api_model.dart';
-import 'package:trus_app/models/api/football/stats/football_all_individual_stats_api_model.dart';
 import 'package:trus_app/models/api/player/player_api_model.dart';
 
-import '../../../config.dart';
+import '../../helper/title_and_text.dart';
 import '../interfaces/json_and_http_converter.dart';
 
 class PlayerSetup implements JsonAndHttpConverter {
   PlayerApiModel? player;
   final List<FootballPlayerApiModel> footballPlayerList;
-  FootballAllIndividualStatsApiModel? playerStats;
+  List<TitleAndText> playerStats;
   final FootballPlayerApiModel primaryFootballPlayer;
   AchievementPlayerDetail? achievementPlayerDetail;
 
@@ -17,7 +16,7 @@ class PlayerSetup implements JsonAndHttpConverter {
     required this.player,
     required this.footballPlayerList,
     required this.primaryFootballPlayer,
-    this.playerStats,
+    required this.playerStats,
     this.achievementPlayerDetail,
   });
 
@@ -36,7 +35,7 @@ class PlayerSetup implements JsonAndHttpConverter {
   factory PlayerSetup.fromJson(Map<String, dynamic> json) {
     return PlayerSetup(
       player: json["player"] != null ? PlayerApiModel.fromJson(json["player"]) : null,
-      playerStats: json["playerStats"] != null ? FootballAllIndividualStatsApiModel.fromJson(json["playerStats"]) : null,
+      playerStats: List<TitleAndText>.from((json['playerStats'] as List<dynamic>).map((titleAndText) => TitleAndText.fromJson(titleAndText))),
       footballPlayerList: List<FootballPlayerApiModel>.from((json['footballPlayerList'] as List<dynamic>).map((footballPlayer) => FootballPlayerApiModel.fromJson(footballPlayer))),
       primaryFootballPlayer: FootballPlayerApiModel.fromJson(json["primaryFootballPlayer"]),
       achievementPlayerDetail: json["achievementPlayerDetail"] != null ? AchievementPlayerDetail.fromJson(json["achievementPlayerDetail"]) : null,
@@ -51,12 +50,5 @@ class PlayerSetup implements JsonAndHttpConverter {
   @override
   String toString() {
     return 'PlayerSetup{player: $player, footballPlayerList: $footballPlayerList, playerStats: $playerStats}';
-  }
-
-  String getPlayerStats() {
-    if(playerStats == null) {
-      return "-";
-    }
-    return "${playerStats!.matches} zápasů / ${playerStats!.goals} gólů";
   }
 }
